+++ pcap4j-core/src/main/java/org/pcap4j/core/Inets.java
@@ -1,6 +1,6 @@
 /*_##########################################################################
   _##
-  _##  Copyright (C) 2011-2012  Pcap4J.org
+  _##  Copyright (C) 2011-2015  Pcap4J.org
   _##
   _##########################################################################
 */
@@ -21,13 +21,58 @@ import com.sun.jna.Platform;
  * @author Kaito Yamada
  * @since pcap4j 0.9.1
  */
-final class Inets {
+public final class Inets {
 
-  static final short AF_UNSPEC = 0;
-  static final short AF_INET;
-  static final short AF_INET6;
-  static final short AF_PACKET;
-  static final short AF_LINK;
+  /**
+   * Unspecified address family.
+   * This value is defined in <code>&lt;sys/socket.h&gt;</code> as 0.
+   */
+  public static final short AF_UNSPEC = 0;
+
+  /**
+   * Address family for IPv4.
+   * This value needs to be the same as AF_INET defined in <code>&lt;sys/socket.h&gt;</code>.
+   * This value may vary depending on OS.
+   * This value is set to 2 by default and can be changed by setting the property
+   * <code>org.pcap4j.af.inet</code> (system property or pcap4j-core.jar/org/pcap4j/pcap4j.properties).
+   *
+   * @see org.pcap4j.Pcap4jPropertiesLoader
+   */
+  public static final short AF_INET;
+
+  /**
+   * Address family for IPv6.
+   * This value needs to be the same as AF_INET6 defined in <code>&lt;sys/socket.h&gt;</code>.
+   * This value varies depending on OS.
+   * By default, this value is set to 30 on Mac OS X, 28 on FreeBSD, 10 on Linux, and 23 on the others.
+   * This value can be changed by setting the property
+   * <code>org.pcap4j.af.inet6</code> (system property or pcap4j-core.jar/org/pcap4j/pcap4j.properties).
+   *
+   * @see org.pcap4j.Pcap4jPropertiesLoader
+   */
+  public static final short AF_INET6;
+
+  /**
+   * Address family for low level packet interface.
+   * This value needs to be the same as AF_PACKET defined in <code>&lt;sys/socket.h&gt;</code>.
+   * This value may vary depending on OS.
+   * This value is set to 17 by default and can be changed by setting the property
+   * <code>org.pcap4j.af.packet</code> (system property or pcap4j-core.jar/org/pcap4j/pcap4j.properties).
+   *
+   * @see org.pcap4j.Pcap4jPropertiesLoader
+   */
+  public static final short AF_PACKET;
+
+  /**
+   * Address family for link layer interface.
+   * This value needs to be the same as AF_LINK defined in <code>&lt;sys/socket.h&gt;</code>.
+   * This value may vary depending on OS.
+   * This value is set to 18 by default and can be changed by setting the property
+   * <code>org.pcap4j.af.link</code> (system property or pcap4j-core.jar/org/pcap4j/pcap4j.properties).
+   *
+   * @see org.pcap4j.Pcap4jPropertiesLoader
+   */
+  public static final short AF_LINK;
 
   static {
     Integer afInet = Pcap4jPropertiesLoader.getInstance().getAfInet();